moved loss in another file
